+++ test/libsolidity/syntaxTests/inlineAssembly/shadowing/no_name_clash_in_import.sol
@@ -9,7 +9,6 @@ contract B {
     function f() public pure {
         assembly {
             let A := 1
-            let A.b := 2
         }
     }
 }